incomplete text box
Incomplete implementation of a text box object that supports wrapping, cropping, scrolling, text cursor, and typing text input. The wrapping is written to run well even when typing into a long document, so long as there are regular line breaks (and thus only a short length of the wrapping needs to be updated).
- Added TextContainer to manage optimal calculations for wrapped/cropped text.
- Added TextBoxState as an entity component to render and manage a text box powered by TextContainer.
- Console will soon use these.
- TextHandler temporarily disabled as I figure out how to properly handle focus changing across various text-input objects and scenarios.